Simulation of central moments
From previous results, I found central moments 1 and 4 showed monotonic
increase along the change of simple 2D patterns.
So I have generated 2D hypnogram images that is as complex as real
hypnogram.
Calculate the central moments of the images of hypnograms and use sort()
function to sort the moments in ascending order.
Using the ordered indices of central moments 1 and 4, I also have
aligned the 2D hypnograms.
The results were not reasonable.